% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_axis_model.R
\name{build_axis_formula}
\alias{build_axis_formula}
\title{Build the model formula for one axis fit}
\usage{
build_axis_formula(iid_effects = NULL)
}
\arguments{
\item{iid_effects}{Character vector of iid effect names.}
}
\value{
A formula object.
}
\description{
Build the model formula for one axis fit
}
\keyword{internal}
