% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_axis_model.R
\name{build_axis_spde}
\alias{build_axis_spde}
\title{Build an SPDE object for one axis model}
\usage{
build_axis_spde(mesh_object, response)
}
\arguments{
\item{mesh_object}{INLA mesh object.}

\item{response}{Numeric response vector.}
}
\value{
An SPDE model object.
}
\description{
Build an SPDE object for one axis model
}
\keyword{internal}
