% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_scores.R
\name{coerce_numeric_column}
\alias{coerce_numeric_column}
\title{Coerce a coordinate column to numeric}
\usage{
coerce_numeric_column(x, column_name)
}
\arguments{
\item{x}{Column vector.}

\item{column_name}{Column name.}
}
\value{
Numeric vector.
}
\description{
Coerce a coordinate column to numeric
}
\keyword{internal}
