% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_axis_surfaces.R
\name{combine_axis_surfaces}
\alias{combine_axis_surfaces}
\title{Combine response-specific surface outputs onto one grid}
\usage{
combine_axis_surfaces(surfaces)
}
\arguments{
\item{surfaces}{Named list of sf surfaces.}
}
\value{
An sf object or \code{NULL}.
}
\description{
Combine response-specific surface outputs onto one grid
}
\keyword{internal}
