% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_status.R
\name{combine_workflow_error_log}
\alias{combine_workflow_error_log}
\title{Combine recorded workflow errors}
\usage{
combine_workflow_error_log(
  dapc_status,
  axis_model_status,
  axis_prediction_status,
  raster_outputs,
  polygon_summaries,
  polygon_metrics
)
}
\description{
Combine recorded workflow errors
}
\keyword{internal}
