% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_status.R
\name{empty_axis_model_status}
\alias{empty_axis_model_status}
\title{Return empty axis-model status}
\usage{
empty_axis_model_status()
}
\description{
Return empty axis-model status
}
\keyword{internal}
