% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_status.R
\name{empty_axis_prediction_status}
\alias{empty_axis_prediction_status}
\title{Return empty axis-prediction status}
\usage{
empty_axis_prediction_status()
}
\description{
Return empty axis-prediction status
}
\keyword{internal}
