% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_status.R
\name{empty_workflow_error_log}
\alias{empty_workflow_error_log}
\title{Return empty workflow error log}
\usage{
empty_workflow_error_log()
}
\description{
Return empty workflow error log
}
\keyword{internal}
