% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{extract_axis_order}
\alias{extract_axis_order}
\title{Extract numeric axis order from column names}
\usage{
extract_axis_order(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Numeric vector.
}
\description{
Extract numeric axis order from column names
}
\keyword{internal}
