% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_inla_mesh.R
\name{extract_boundary_coordinates}
\alias{extract_boundary_coordinates}
\title{Extract polygon boundary coordinates for mesh construction}
\usage{
extract_boundary_coordinates(boundary_sf)
}
\arguments{
\item{boundary_sf}{An sf boundary object.}
}
\value{
A two-column numeric matrix.
}
\description{
Extract polygon boundary coordinates for mesh construction
}
\keyword{internal}
