% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_metrics.R
\name{extract_polygon_metric_columns}
\alias{extract_polygon_metric_columns}
\title{Extract selected metric columns with missing columns represented by NA}
\usage{
extract_polygon_metric_columns(data, columns)
}
\arguments{
\item{data}{Polygon summary data frame.}

\item{columns}{Requested columns.}
}
\value{
Numeric matrix.
}
\description{
Extract selected metric columns with missing columns represented by NA
}
\keyword{internal}
