% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{extract_variant_matrix}
\alias{extract_variant_matrix}
\title{Extract a variant matrix from supported inputs}
\usage{
extract_variant_matrix(x)
}
\arguments{
\item{x}{A matrix or list containing \code{variant_matrix}.}
}
\value{
A numeric matrix.
}
\description{
Extract a variant matrix from supported inputs
}
\keyword{internal}
