% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{join_scores_metadata}
\alias{join_scores_metadata}
\title{Join score tables to metadata}
\usage{
join_scores_metadata(scores, metadata = NULL, id_col = "sequence_id")
}
\arguments{
\item{scores}{A data frame with a \code{sequence_id} column.}

\item{metadata}{Optional metadata data frame.}

\item{id_col}{Name of the sequence identifier column in \code{metadata}.}
}
\value{
A data frame.
}
\description{
Join score tables to metadata
}
\keyword{internal}
