% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_inla_mesh.R
\name{normalize_boundary}
\alias{normalize_boundary}
\title{Normalize a boundary input to sf}
\usage{
normalize_boundary(boundary)
}
\arguments{
\item{boundary}{A boundary object or path.}
}
\value{
An sf object.
}
\description{
Normalize a boundary input to sf
}
\keyword{internal}
