% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_axis_model.R
\name{normalize_inla_mesh}
\alias{normalize_inla_mesh}
\title{Normalize a mesh input for axis modeling}
\usage{
normalize_inla_mesh(mesh)
}
\arguments{
\item{mesh}{A mesh object or mesh list.}
}
\value{
An INLA mesh object.
}
\description{
Normalize a mesh input for axis modeling
}
\keyword{internal}
