% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{normalize_workflow_alignment}
\alias{normalize_workflow_alignment}
\title{Normalize workflow alignment input for encoding}
\usage{
normalize_workflow_alignment(alignment)
}
\arguments{
\item{alignment}{Alignment object returned by \code{\link[=read_alignment]{read_alignment()}}.}
}
\value{
A DNAbin alignment.
}
\description{
Normalize workflow alignment input for encoding
}
\keyword{internal}
