% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{prepare_prediction_surface_table}
\alias{prepare_prediction_surface_table}
\title{Prepare prediction surfaces for CSV output}
\usage{
prepare_prediction_surface_table(combined_surface, crs_input, lon_col, lat_col)
}
\arguments{
\item{combined_surface}{Combined prediction sf object.}

\item{crs_input}{Input CRS.}

\item{lon_col}{Longitude column name.}

\item{lat_col}{Latitude column name.}
}
\value{
Data frame.
}
\description{
Prepare prediction surfaces for CSV output
}
\keyword{internal}
