% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_outputs.R
\name{resolve_axis_polygons}
\alias{resolve_axis_polygons}
\title{Resolve polygon inputs}
\usage{
resolve_axis_polygons(polygons)
}
\arguments{
\item{polygons}{Sf polygons or a path.}
}
\value{
An sf object.
}
\description{
Resolve polygon inputs
}
\keyword{internal}
