% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_outputs.R
\name{resolve_axis_raster}
\alias{resolve_axis_raster}
\title{Resolve a raster input}
\usage{
resolve_axis_raster(raster)
}
\arguments{
\item{raster}{Raster or path.}
}
\value{
A terra SpatRaster.
}
\description{
Resolve a raster input
}
\keyword{internal}
