% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_metrics.R
\name{row_mean_or_na}
\alias{row_mean_or_na}
\title{Calculate a row-wise mean with all-missing rows as NA}
\usage{
row_mean_or_na(values)
}
\arguments{
\item{values}{Numeric matrix.}
}
\value{
Numeric vector.
}
\description{
Calculate a row-wise mean with all-missing rows as NA
}
\keyword{internal}
