% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_metrics.R
\name{row_sum_of_squares}
\alias{row_sum_of_squares}
\title{Calculate a row-wise sum of squares with all-missing rows as NA}
\usage{
row_sum_of_squares(values)
}
\arguments{
\item{values}{Numeric matrix.}
}
\value{
Numeric vector.
}
\description{
Calculate a row-wise sum of squares with all-missing rows as NA
}
\keyword{internal}
