% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_status.R
\name{scalar_character}
\alias{scalar_character}
\title{Normalize optional scalar to character}
\usage{
scalar_character(value)
}
\description{
Normalize optional scalar to character
}
\keyword{internal}
