% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencemap-package.R
\docType{package}
\name{sequencemap-package}
\alias{sequencemap}
\alias{sequencemap-package}
\title{sequencemap: Sequence Mapping and Spatial Analysis Tools}
\description{
Package infrastructure for phylogenetic, population genetic, and spatial
epidemiology workflows.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
