% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_scores.R
\name{summarize_extra_column}
\alias{summarize_extra_column}
\title{Summarize an auxiliary metadata column at location level}
\usage{
summarize_extra_column(x)
}
\arguments{
\item{x}{Vector of values.}
}
\value{
A scalar summary value.
}
\description{
Summarize an auxiliary metadata column at location level
}
\keyword{internal}
