% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_scores.R
\name{summarize_location_coordinates}
\alias{summarize_location_coordinates}
\title{Summarize location coordinates and identifiers}
\usage{
summarize_location_coordinates(group_data, location_col, lon_col, lat_col)
}
\arguments{
\item{group_data}{Grouped data frame.}

\item{location_col}{Optional location column.}

\item{lon_col}{Longitude column.}

\item{lat_col}{Latitude column.}
}
\value{
Named list.
}
\description{
Summarize location coordinates and identifiers
}
\keyword{internal}
