% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_scores.R
\name{transform_div_trace}
\alias{transform_div_trace}
\title{Transform a diversity trace statistic}
\usage{
transform_div_trace(x, transform)
}
\arguments{
\item{x}{Numeric scalar.}

\item{transform}{Transformation name.}
}
\value{
Numeric scalar.
}
\description{
Transform a diversity trace statistic
}
\keyword{internal}
