% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_outputs.R
\name{validate_axis_surface_points}
\alias{validate_axis_surface_points}
\title{Validate a projected prediction surface}
\usage{
validate_axis_surface_points(surface)
}
\arguments{
\item{surface}{An sf point surface.}
}
\value{
TRUE, invisibly.
}
\description{
Validate a projected prediction surface
}
\keyword{internal}
