% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_metrics.R
\name{validate_polygon_metric_options}
\alias{validate_polygon_metric_options}
\title{Validate options for polygon metrics}
\usage{
validate_polygon_metric_options(
  axis_prefixes,
  axis_numbers,
  pred_stat,
  heterogeneity_stat,
  uncertainty_stat,
  polygon_id_col
)
}
\arguments{
\item{axis_prefixes}{Optional prefixes.}

\item{axis_numbers}{Optional axis numbers.}

\item{pred_stat}{Prediction statistic.}

\item{heterogeneity_stat}{Heterogeneity statistic.}

\item{uncertainty_stat}{Uncertainty statistic.}

\item{polygon_id_col}{Optional identifier column.}
}
\value{
TRUE, invisibly.
}
\description{
Validate options for polygon metrics
}
\keyword{internal}
