% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_inla_mesh.R
\name{validate_projected_sf_points}
\alias{validate_projected_sf_points}
\title{Validate projected sf point input}
\usage{
validate_projected_sf_points(x, object_name = "x")
}
\arguments{
\item{x}{An sf object.}

\item{object_name}{Name used in error messages.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Validate projected sf point input
}
\keyword{internal}
