% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{validate_variant_matrix}
\alias{validate_variant_matrix}
\title{Validate a variant matrix before ordination}
\usage{
validate_variant_matrix(variant_matrix)
}
\arguments{
\item{variant_matrix}{A numeric matrix.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Validate a variant matrix before ordination
}
\keyword{internal}
