% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_status.R
\name{workflow_error_row}
\alias{workflow_error_row}
\title{Create one workflow error row}
\usage{
workflow_error_row(stage, response_col, message, class)
}
\description{
Create one workflow error row
}
\keyword{internal}
