% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_message}
\alias{workflow_message}
\title{Emit a workflow message when verbose is enabled}
\usage{
workflow_message(verbose, text)
}
\arguments{
\item{verbose}{Logical flag.}

\item{text}{Message text.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Emit a workflow message when verbose is enabled
}
\keyword{internal}
